add testcontainers, test configuration, add it tests, start implementing logic

+++ src/main/resources/db.changelog/db.changelog-2024-01_init_schemas.sql
@@ -16,7 +16,7 @@ CREATE TABLE movie
 (
     id               VARCHAR(255)  NOT NULL,
     name             VARCHAR(255)  NOT NULL,
-    is_winner        TINYINT       NOT NULL,
+    is_winner        INTEGER       NOT NULL,
     office_box_value DECIMAL(10,3),
 
     CONSTRAINT pk_movie PRIMARY KEY (id)
